refactor: Rename cmd packange to command
