+++ src/main/java/InsuranceSystem.java
@@ -1,4 +1,4 @@
-import cmd.AuthCommand;
+import command.AuthCommand;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import utils.Session;
